Convert rel to calc as f64

+++ crate/src/exec/basic/count.rs
@@ -2,7 +2,7 @@ use std::{collections::HashSet, sync::Arc};
 
 use arrow::{array::{Array, Float64Array, StringArray}, datatypes::{DataType, Field, Schema}, record_batch::RecordBatch};
 
-use crate::{exec::basic::analyze::{filter_with, find_unique_string}, get_record_batches};
+use crate::{exec::basic::analyze::{filter_with, find_unique_string}, get_record_batches, utils::print_to_cons::print_to_js_with_obj};
 
 use super::analyze;
 
@@ -143,7 +143,7 @@ pub fn relative(batch: &RecordBatch, column_index_for_numerator: usize, column_i
 
     let numerator = vec_nom.len();
 
-    let result = (numerator / denominator) as f64; 
+    let result = numerator as f64 / denominator as f64; 
 
     let mut result_builder = Float64Array::builder(1);
